+++ YooKassa/Message/AbstractRequest.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Yandex.Kassa driver for Omnipay payment processing library
+ * YooKassa driver for Omnipay payment processing library
  *
- * @link      https://github.com/hiqdev/omnipay-yandex-kassa
- * @package   omnipay-yandex-kassa
+ * @link      https://github.com/igor-tv/omnipay-yookassa
+ * @package   omnipay-yookassa
  * @license   MIT
- * @copyright Copyright (c) 2019, HiQDev (http://hiqdev.com/)
+ * @copyright Copyright (c) 2021, Igor Tverdokhleb, igor-tv@mail.ru
  */
 
-namespace Omnipay\YandexKassa\Message;
+namespace Omnipay\YooKassa\Message;
 
-use YandexCheckout\Client;
+use YooKassaCheckout\Client;
 
 /**
  * Class AbstractRequest.
@@ -44,7 +44,7 @@ abstract class AbstractRequest extends \Omnipay\Common\Message\AbstractRequest
         return $this->setParameter('secret', $value);
     }
 
-    public function setYandexClient(Client $client): void
+    public function setYooKassaClient(Client $client): void
     {
         $this->client = $client;
     }