+++ tests/unit/GatewayTest.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * Yandex.Kassa driver for Omnipay payment processing library
+ * YooKassa driver for Omnipay payment processing library
  *
- * @link      https://github.com/hiqdev/omnipay-yandex-kassa
- * @package   omnipay-yandex-kassa
+ * @link      https://github.com/igor-tv/omnipay-yookassa
+ * @package   omnipay-yookassa
  * @license   MIT
- * @copyright Copyright (c) 2019, HiQDev (http://hiqdev.com/)
+ * @copyright Copyright (c) 2021, Igor Tverdokhleb, igor-tv@mail.ru
  */
 
-namespace Omnipay\YandexKassa\Tests;
+namespace Omnipay\YooKassa\Tests;
 
 use Omnipay\Tests\GatewayTestCase;
-use Omnipay\YandexKassa\Gateway;
+use Omnipay\YooKassa\Gateway;
 
 class GatewayTest extends GatewayTestCase
 {