+++ tests/unit/Message/PurchaseRequestTest.php
@@ -1,21 +1,21 @@
 <?php
 /**
- * Yandex.Kassa driver for Omnipay payment processing library
+ * YooKassa driver for Omnipay payment processing library
  *
- * @link      https://github.com/hiqdev/omnipay-yandex-kassa
- * @package   omnipay-yandex-kassa
+ * @link      https://github.com/igor-tv/omnipay-yookassa
+ * @package   omnipay-yookassa
  * @license   MIT
- * @copyright Copyright (c) 2019, HiQDev (http://hiqdev.com/)
+ * @copyright Copyright (c) 2021, Igor Tverdokhleb, igor-tv@mail.ru
  */
 
-namespace Omnipay\YandexKassa\Tests\Message;
+namespace Omnipay\YooKassa\Tests\Message;
 
-use Omnipay\YandexKassa\Message\PurchaseRequest;
-use Omnipay\YandexKassa\Message\PurchaseResponse;
+use Omnipay\YooKassa\Message\PurchaseRequest;
+use Omnipay\YooKassa\Message\PurchaseResponse;
 
 class PurchaseRequestTest extends TestCase
 {
-    /** @var \Omnipay\YandexKassa\Message\PurchaseRequest */
+    /** @var \Omnipay\YooKassa\Message\PurchaseRequest */
     private $request;
 
     private $shopId         = '54401';
@@ -33,7 +33,7 @@ class PurchaseRequestTest extends TestCase
 
         $this->request = new PurchaseRequest($this->getHttpClient(), $this->getHttpRequest());
         $this->request->initialize([
-            'yandexClient'  => $this->buildYandexClient($this->shopId, $this->secretKey),
+            'yooKassaClient'  => $this->buildYooKassaClient($this->shopId, $this->secretKey),
             'transactionId' => $this->transactionId,
             'amount'        => $this->amount,
             'currency'      => $this->currency,
@@ -62,7 +62,7 @@ class PurchaseRequestTest extends TestCase
             ['http_code' => 200],
         ]);
 
-        $this->getYandexClient($this->request)
+        $this->getYooKassaClient($this->request)
              ->setApiClient($curlClientStub)
              ->setAuth($this->shopId, $this->secretKey);