+++ tests/unit/Message/PurchaseResponseTest.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * Yandex.Kassa driver for Omnipay payment processing library
+ * YooKassa driver for Omnipay payment processing library
  *
- * @link      https://github.com/hiqdev/omnipay-yandex-kassa
- * @package   omnipay-yandex-kassa
+ * @link      https://github.com/igor-tv/omnipay-yookassa
+ * @package   omnipay-yookassa
  * @license   MIT
- * @copyright Copyright (c) 2019, HiQDev (http://hiqdev.com/)
+ * @copyright Copyright (c) 2021, Igor Tverdokhleb, igor-tv@mail.ru
  */
 
-namespace Omnipay\YandexKassa\Tests\Message;
+namespace Omnipay\YooKassa\Tests\Message;
 
-use Omnipay\YandexKassa\Message\PurchaseRequest;
-use Omnipay\YandexKassa\Message\PurchaseResponse;
+use Omnipay\YooKassa\Message\PurchaseRequest;
+use Omnipay\YooKassa\Message\PurchaseResponse;
 
 class PurchaseResponseTest extends TestCase
 {
@@ -33,7 +33,7 @@ class PurchaseResponseTest extends TestCase
 
         $this->request = new PurchaseRequest($this->getHttpClient(), $this->getHttpRequest());
         $this->request->initialize([
-            'yandexClient'  => $this->buildYandexClient($this->shopId, $this->secretKey),
+            'yooKassaClient'  => $this->buildYooKassaClient($this->shopId, $this->secretKey),
             'transactionId' => $this->transactionId,
             'amount'        => $this->amount,
             'currency'      => $this->currency,
@@ -51,7 +51,7 @@ class PurchaseResponseTest extends TestCase
             ['http_code' => 200],
         ]);
 
-        $this->getYandexClient($this->request)
+        $this->getYooKassaClient($this->request)
              ->setApiClient($curlClientStub)
              ->setAuth($this->shopId, $this->secretKey);
 
@@ -63,7 +63,7 @@ class PurchaseResponseTest extends TestCase
         $this->assertNull($response->getCode());
         $this->assertNull($response->getMessage());
         $this->assertSame('GET', $response->getRedirectMethod());
-        $this->assertSame("https://money.yandex.ru/api-pages/v2/payment-confirm/epl?orderId={$response->getTransactionReference()}", $response->getRedirectUrl());
+        $this->assertSame("https://yoomoney.ru/api-pages/v2/payment-confirm/epl?orderId={$response->getTransactionReference()}", $response->getRedirectUrl());
         $this->assertSame($this->transactionId, $response->getTransactionId());
         $this->assertSame('247732b9-000f-5000-a000-13d9c7c381a8', $response->getTransactionReference());
         $this->assertEmpty($response->getRedirectData());