+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Under development]
 
+## [4.0.0] - 2021-05-10
+
+- Full rebranding from Yandex.Kassa to YooKassa ([@igor-tv]) 
+- Migrating to YooKassa API PHP Client Library. All gateway url's was changed.
+
 ## [3.0.2] - 2019-08-06
 
 - Fixed `DetailsResponse::getPayer()`: Payment method title might be missing, so a fallback was implemented ([@SilverFire])
@@ -24,8 +29,9 @@
 [andreyklochok@gmail.com]: https://github.com/tafid
 [@BladeRoot]: https://github.com/BladeRoot
 [bladeroot@gmail.com]: https://github.com/BladeRoot
+[@igor-tv]: https://github.com/igor-tv
 [Under development]: https://github.com/hiqdev/omnipay-yandex-kassa/compare/3.0.1...HEAD
 [3.0.0]: https://github.com/hiqdev/omnipay-yandex-kassa/releases/tag/3.0.0
-[Development started]: https://github.com/hiqdev/omnipay-yandex-kassa/compare/2.0.0...Development started
 [3.0.1]: https://github.com/hiqdev/omnipay-yandex-kassa/compare/3.0.0...3.0.1
 [3.0.2]: https://github.com/hiqdev/omnipay-yandex-kassa/compare/3.0.1...3.0.2
+[4.0.0]: https://github.com/igor-tv/omnipay-yookassa/compare/3.0.2...4.0.0